Update EchoClient.java by Joining Starter and Reference Code
In this commit, I synthesize the last lab's code I brought in for reference with the starter code for this lab. To do this, I set the server to being "localhost" at all times and split the main method into another helper method: start(). This will make way for the creation of threads.